refactor: use PostCard instead of Post and PostCard
