--- client/src/pages/Post.tsx
@@ -1,34 +1,0 @@
-import React from 'react';
-import { useNavigate } from 'react-router-dom';
-import styles from '../styles/Post.module.css';
-
-type PostProps = {
-  id: string;
-  username: string;
-  profilePic: string;
-  title: string;
-  description: string;
-  image: string;
-};
-
-const Post: React.FC<PostProps> = ({ id, username, profilePic, title, description, image }) => {
-  const navigate = useNavigate();
-
-  const handleClick = () => {
-    navigate(`/posts/${id}`);
-  };
-
-  return (
-    <div className={styles.postContainer} onClick={handleClick}>
-      <div className={styles.header}>
-        <img src={profilePic} alt={`${username}'s profile`} className={styles.profilePic} />
-        <span className={styles.username}>{username}</span>
-      </div>
-      <h3 className={styles.title}>{title}</h3>
-      <p className={styles.description}>{description}</p>
-      <img src={image} alt={title} className={styles.postImage} />
-    </div>
-  );
-};
-
-export default Post;
