+++ client/src/pages/PostDetail.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import styles from '../styles/PostDetail.module.css';
